# Trucking-Game-API/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Trucking Software Public API",
		"description": "API for trucking software web application",
		"version": "1.0.0"
	},
	"servers": [
		{
			"url": "http://localhost:3001/api/v1",
			"description": "Development server"
		}
	],
	"paths": {
...
		"/articles/search": {
			"get": {
				"description": "Get list of articles that match the provided query",
				"summary": "Get list of articles",
				"operationId": "getArticlesByQuery",
				"tags": ["articles"],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"schema": {
							"type": "integer"
						},
						"required": false,
						"description": "The limit of documents to return. Defaults to 10."
					},
					{
						"in": "query",
						"name": "page",
						"schema": {
							"type": "integer"
						},
						"required": false,
						"description": "The page of documents to return, will determine the amount of documents to skip when querying. Defaults to 1."
					},
					{
						"in": "query",
						"name": "tags",
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						},
						"required": false,
						"description": "Array of tags to execute a search for"
					},
					{
						"in": "query",
						"name": "authorId",
						"schema": {
							"type": "string"
						},
						"required": false,
						"description": "Id of the author to execute a search for"
					}
				],
				"responses": {
					"200": {
						"description": "Query articles successful response",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"articles": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/Article"
											}
										},
										"totalDocuments": {
											"type": "integer"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Query articles unsuccessful response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
		"/articles/{articleSlug}/comments": {
			"get": {
				"description": "Get commendts for a specific article identified by a slug",
				"summary": "Get comments for a specific article",
				"operationId": "getCommentsByArticle",
				"tags": ["articles"],
				"parameters": [
					{
						"in": "path",
						"name": "articleSlug",
						"schema": {
							"type": "string"
						},
						"required": true,
						"description": "Slug for the article that you want to find comments for"
					}
				],
				"responses": {
					"200": {
						"description": "Get Article comments successful response",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": "true"
										},
										"comments": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/ArticleComment"
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Get Article comments unsuccessful response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
		"/articles/{articleId}": {
			"delete": {
				"description": "Delete a single article by id",
				"summary": "Delete a single article",
				"operationId": "deleteOneArticle",
				"tags": ["articles"],
				"parameters": [
					{
						"in": "path",
						"name": "articleId",
						"schema": {
							"type": "string"
						},
						"required": true,
						"description": "Id of the article to delete"
					}
				],
				"responses": {
					"200": {
						"description": "Delete one article successful response",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": "true"
										},
										"message": {
											"type": "string",
											"example": "Successfully deleted article"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Delete one article unsuccessful response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			},
			"patch": {
				"description": "Update a single article by id",
				"summary": "Update a single article",
				"operationId": "updateArticle",
				"tags": ["articles"],
				"parameters": [
					{
						"in": "path",
						"name": "articleId",
						"schema": {
							"type": "string"
						},
						"required": true,
						"description": "Id of the article to update"
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ArticleBody"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Update article successful response",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": "true"
										},
										"updatedArticle": {
											"$ref": "#/components/schemas/Article"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Update article unsuccessful response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
		"/articles/multiple": {
			"delete": {
				"description": "Delete multiple articles by id",
				"summary": "Delete multiple articles",
				"operationId": "deleteMultipleArticles",
				"tags": ["articles"],
				"parameters": [
					{
						"in": "query",
						"name": "articleId",
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						},
						"required": true,
						"description": "Id(s) of article(s) to delete"
					}
				],
				"responses": {
					"200": {
						"description": "Delete multiple articles successful response",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": "true"
										},
										"message": {
											"type": "string",
											"example": "Successfully deleted articles"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Delete multiple articles unsuccessful response",
						"content": {
							"application/json": {
								"schema": {
									"oneOf": [
										{
											"$ref": "#/components/schemas/ErrorResponse"
										},
										{
											"$ref": "#/components/schemas/DeleteMultipleErrorResponse"
										}
									]
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/articles/toggleEditing/{articleId}": {
			"patch": {
				"description": "Toggle editing on a specific article",
				"summary": "Toggle editing on a specific article",
				"operationId": "toggleEditingLocked",
				"tags": ["articles"],
				"parameters": [
					{
						"in": "path",
						"name": "articleId",
						"schema": {
							"type": "string"
						},
						"required": true,
						"description": "Id of the article to update"
					}
				],
				"responses": {
					"200": {
						"description": "Toggle editing successful response",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": "true"
										},
										"article": {
											"$ref": "#/components/schemas/Article"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Toggle editing unsuccessful response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},
		"schemas": {
			"Article": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"author": {
						"$ref": "#/components/schemas/User"
					},
					"slug": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"htmlBody": {
						"type": "string"
					},
					"textBody": {
						"type": "string"
					},
					"tags": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"comments": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ArticleComment"
						}
					},
					"isFree": {
						"type": "boolean"
					},
					"editingLocked": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string"
					}
				}
			},
			"ArticleBody": {
				"type": "object",
				"properties": {
					"title": {
						"type": "string"
					},
					"htmlBody": {
						"type": "string"
					},
					"textBody": {
						"type": "string"
					},
					"tags": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"category": {
						"type": "string"
					},
					"isFree": {
						"type": "boolean"
					}
				},
				"required": ["title", "htmlBody", "textBody", "category", "isFree"]
			},
			"ArticleComment": {
				"type": "object",
				"properties": {
					"text": {
						"type": "string"
					},
					"relatedArticle": {
						"$ref": "#/components/schemas/Article"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"createdAt": {
						"type": "string"
					}
				}
			},
...
			"DeleteMultipleErrorResponse": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean",
						"example": "false"
					},
					"message": {
						"type": "string",
						"example": "One or more artcles was not found"
					},
					"incomplete": {
						"type": "array",
						"items": {
							"type": "string"
						}
					}
				}
			},
			"ErrorResponse": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean",
						"example": "false"
					},
					"error": {
						"type": "string"
					}
				}
			},
...
			"User": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"type": "integer"
					},
					"phone": {
						"type": "string"
					},
					"profileImage": {
						"type": "string"
					},
					"hasPaid": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string"
					}
				}
			},
...
		}
	},
	"security": [
		{
			"bearerAuth": []
		}
	]
}